Fixing a bug when growing the stack while calling a class method.
git-svn-id: http://svn.code.sf.net/p/angelscript/code/trunk@573 404ce1b2-830e-0410-a2e2-b09542c77caf

+++ sdk/angelscript/source/as_restore.cpp
@@ -1443,7 +1443,7 @@ void asCRestore::TranslateFunction(asCScriptFunction *func)
 		{
 			// Translate the index to the func pointer
 			asPTRWORD *fid = (asPTRWORD*)&bc[n+1];
-			*fid = (asPTRWORD)(size_t)FindFunction(*fid);
+			*fid = (asPTRWORD)(size_t)FindFunction((int)*fid);
 		}
 		else if( c == asBC_ALLOC )
 		{
